# [elastic] Parse the whole workspace folder before the code intelligence