Parse the whole workspace folder before handle the lsp requests. The
parse is not time-consuming, in contrast, pre-parse the workspace folder
will reduce the 'go list' invokes. 'go list' was invoked for any
sub-folders originally, langserver invokes 'go list' just for the whole
workspace folder for now.

For now, langserver doesn't support '_test.go' files, for the '_test.go'
files, langserver will invoke the 'go list' as usual. That's why we still
keep the go list cache. Once langserver supports '_test.go' files,
redesign the go list cache.

Besides that, parse will fail for the whole workspace folder when we
disable the deps download because of the broken dependencies.