@movie-travel-code 2019-10-21
| repo | Symbol Count(pred) | Symbol Count(current) | Time(pred) | Time(current) | File Count(pred) | File Cound(current) |
|:--|:--|:--|:--|:--|:--|:--|
| tidb | 26308 | 26342 | 161.438s | 113.695s | 1035 | 1035 |
| mongo-go-driver | 7152 | 7152 | 110.991s | 100.913s | 2283 | 2283 |
| apm-agent-go | 2639 | 2639 | 66.949s | 61.774s | 441 | 441 |

The pre-parse for the whole workspace folder improves the performance to some extent. The bigger repo can get greater performance improvement.